refactor: separate tracing-specific functionality
