+++ src/tracing/index.ts
@@ -18,15 +18,15 @@ import { context, propagation, trace } from '@opentelemetry/api';
 import { NodeTracerProvider } from '@opentelemetry/sdk-trace-node';
 import { registerInstrumentations } from '@opentelemetry/instrumentation';
 
-import { configureHttpInstrumentation } from './instrumentations/http';
-import { configureLogInjection } from './instrumentations/logging';
+import { configureHttpInstrumentation } from '../instrumentations/http';
+import { configureLogInjection } from '../instrumentations/logging';
 import { Options, _setDefaultOptions } from './options';
 import { gte } from 'semver';
 import {
   AsyncHooksContextManager,
   AsyncLocalStorageContextManager,
 } from '@opentelemetry/context-async-hooks';
-import { configureRedisInstrumentation } from './instrumentations/redis';
+import { configureRedisInstrumentation } from '../instrumentations/redis';
 
 let unregisterInstrumentations: (() => void) | null = null;
