+++ test/propagation.test.ts
@@ -30,8 +30,8 @@ import {
   InMemorySpanExporter,
   SpanProcessor,
 } from '@opentelemetry/sdk-trace-base';
-import { SYNTHETIC_RUN_ID_FIELD } from '../src/SplunkBatchSpanProcessor';
-import { defaultSpanProcessorFactory } from '../src/options';
+import { SYNTHETIC_RUN_ID_FIELD } from '../src/tracing/SplunkBatchSpanProcessor';
+import { defaultSpanProcessorFactory } from '../src/tracing/options';
 import * as utils from './utils';
 
 function assertIncludes(arr: string[], item: string) {
